# docs/DOC-05-api-spec.yaml
openapi: 3.1.0
info:
  title: Contract Intelligence API
  version: 0.3.0
  status: enterprise-ready — đã chốt kiến trúc Bearer JWT, x-rbac & Tenant Isolation (Sprint 1)
  description: |
    REST API cho Contract Intelligence — hệ thống trích xuất, so sánh và review hợp đồng cấp Enterprise.

    **Khớp kiến trúc:** `docs/DOC-04-architecture.md` (v0.1, 2026-09-16)
    **Khớp schema:** `docs/DOC-04b-postgres-schema.sql` (PostgreSQL v1.0 đã chốt 2026-09-17)

    ## Cơ chế Xác thực & Phân quyền Enterprise
    - **Xác thực (Bearer JWT):** Mọi request (trừ probe `/healthz`, `/readyz`) bắt buộc mang header:
      `Authorization: Bearer <jwt_token>`
      JWT payload mang các claims: `sub` (user_id), `tenant_id`, `role` (`OPERATOR` | `REVIEWER` | `ADMINISTRATOR`).
    - **Tenant Isolation (`X-Tenant-Id`):** Client bắt buộc gửi header `X-Tenant-Id`.
      Backend enforce tenant boundary nghiêm ngặt: `X-Tenant-Id` phải trùng khớp với claim `tenant_id` trong JWT. Mọi thao tác ghi/đọc dữ liệu đều bị cô lập theo tenant.
    - **Phân quyền Role-Based Access Control (`x-rbac`):** Mỗi endpoint quy định chặt chẽ danh sách role hợp lệ qua extension property `x-rbac`:
      - `OPERATOR`: Khởi tạo dossier/batch, upload văn bản, xác nhận manifest, kích hoạt run, tạo yêu cầu Re-OCR.
      - `REVIEWER`: Xem dossier/documents/facts/conflicts/citations, thực hiện review actions, xem lịch sử revision, đề xuất khóa dossier.
...

    ## Đơn vị xử lý & Nghiệp vụ cốt lõi (9 Phân hệ Enterprise)
...

...

...

  contact:
...
servers:
...
paths:
...
  /dossiers/{dossier_id}/runs:
    parameters:
      - $ref: '#/components/parameters/DossierId'
    post:
      tags: [runs]
      summary: Khởi tạo pipeline execution run mới cho dossier
      description: Tạo snapshot thực thi mới với config override tuỳ chọn (prompt version, OCR profile, confidence threshold).
      operationId: triggerDossierRun
      x-rbac: [OPERATOR, ADMINISTRATOR]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRunRequest'
      responses:
        '202':
          description: Pipeline run accepted and queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelopePipelineRun'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Pipeline run conflict (run currently active)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  /dossiers/{dossier_id}/reprocess:
    parameters:
      - $ref: '#/components/parameters/DossierId'
    post:
      tags: [dossiers]
      summary: Tạo job mới cho dossier (chạy lại với cấu hình hiện tại)
      description: |
        Tạo job mới + `pipeline_run` mới, kết quả cũ vẫn còn (bất biến).
        Khi job mới chạy xong, UI sẽ hiển thị kết quả mới nhất; kết quả cũ có thể xem qua `/dossiers/{id}/audit`.
      operationId: reprocessDossier
      x-rbac: [OPERATOR, ADMINISTRATOR]
      responses:
        '202':
          description: Reprocess accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelopeDossierCreated'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Dossier đang có job chạy, không thể reprocess
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'

  # ─── Runs ────────────────────────────────────────────────────────────────
  /runs:
    get:
      tags: [runs]
      summary: Danh sách tất cả các pipeline execution runs trong tenant
      operationId: listRuns
      x-rbac: [OPERATOR, REVIEWER, ADMINISTRATOR]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: dossier_id
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [queued, running, completed, failed, cancelled] }
      responses:
        '200':
          description: List of pipeline runs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelopePipelineRunList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /runs/{run_id}:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      tags: [runs]
      summary: Chi tiết snapshot pipeline run (metrics, stage tracking, config snapshot, logs)
      operationId: getRunDetail
      x-rbac: [OPERATOR, REVIEWER, ADMINISTRATOR]
      responses:
        '200':
          description: Run detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelopePipelineRunDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /runs/{run_id}/cancel:
    parameters:
      - $ref: '#/components/parameters/RunId'
    post:
      tags: [runs]
      summary: Huỷ bỏ pipeline run đang chạy
      operationId: cancelRun
      x-rbac: [OPERATOR, ADMINISTRATOR]
      responses:
        '200':
          description: Pipeline run cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelopePipelineRun'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Run cannot be cancelled (already completed or failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'

  # ─── Jobs ────────────────────────────────────────────────────────────────
...
  /documents/{document_id}/pages:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    get:
      tags: [pages]
      summary: Metadata các trang (loại, kích thước, rotation, chỉ số chất lượng)
      operationId: listPages
      x-rbac: [OPERATOR, REVIEWER, ADMINISTRATOR]
      responses:
        '200':
          description: Page metadata list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelopePageList'
  /documents/{document_id}/pages/{page_no}/image:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
      - name: page_no
        in: path
        required: true
        schema: { type: integer, minimum: 1 }
      - name: variant
        in: query
        schema:
          type: string
          enum: [preview, render]
          default: preview
    get:
      tags: [pages]
      summary: Ảnh trang (PNG 300 DPI hoặc WebP preview)
      operationId: getPageImage
      x-rbac: [OPERATOR, REVIEWER, ADMINISTRATOR]
      responses:
        '200':
          description: Page image
          content:
            image/png: {}
            image/webp: {}
        '404':
          $ref: '#/components/responses/NotFound'
...
  /documents/{document_id}/clauses:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
      - name: run_id
        in: query
        schema: { type: string }
    get:
      tags: [clauses]
      summary: Cây điều khoản (Điều → Khoản → Điểm) + region bbox
      operationId: getClauseTree
      x-rbac: [OPERATOR, REVIEWER, ADMINISTRATOR]
      responses:
        '200':
          description: Clause tree
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelopeClauseTree'
  /documents/{document_id}/tables:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
      - name: run_id
        in: query
        schema: { type: string }
    get:
      tags: [tables]
      summary: Bảng có cấu trúc (không flatten)
      operationId: getTables
      x-rbac: [OPERATOR, REVIEWER, ADMINISTRATOR]
      responses:
        '200':
          description: Tables
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelopeTableList'

  # ─── Facts ───────────────────────────────────────────────────────────────
  /dossiers/{dossier_id}/facts:
    parameters:
      - $ref: '#/components/parameters/DossierId'
      - name: effective
        in: query
        description: Nếu true, trả kèm `current_version` cho optimistic concurrency
        schema: { type: boolean, default: true }
      - name: key
        in: query
        description: Lọc theo fact key (vd. `price.total`)
        schema: { type: string }
    get:
      tags: [facts]
      summary: Danh sách fact + giá trị hiệu lực (kèm version cho concurrency)
      operationId: listFacts
      x-rbac: [OPERATOR, REVIEWER, ADMINISTRATOR]
      responses:
        '200':
          description: Fact list with effective values
          headers:
            ETag:
              description: ETag cho optimistic concurrency ở cấp dossier
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelopeFactList'

  # ─── Findings & Conflicts ────────────────────────────────────────────────
...
  /citations/{citation_id}:
    parameters:
      - name: citation_id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [citations]
      summary: Resolve citation — trả URL ảnh trang + bbox để UI highlight một thao tác (BR-07)
      operationId: getCitation
      x-rbac: [OPERATOR, REVIEWER, ADMINISTRATOR]
      responses:
        '200':
          description: Citation resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelopeCitationResolved'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── Review ──────────────────────────────────────────────────────────────
...
components:
...
  parameters:
...
    RunId:
      name: run_id
      in: path
      required: true
      schema: { type: string, pattern: "^run_[0-9A-HJKMNP-TV-Z]{26}$" }
...
    DossierId:
      name: dossier_id
      in: path
      required: true
      schema: { type: string, pattern: "^dos_[0-9A-HJKMNP-TV-Z]{26}$" }
...
    DocumentId:
      name: document_id
      in: path
      required: true
      schema: { type: string, pattern: "^doc_[0-9A-HJKMNP-TV-Z]{26}$" }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }

  # ── Schemas ──────────────────────────────────────────────────────────────
  schemas:
...
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri, default: "about:blank" }
        title: { type: string }
        status: { type: integer }
        code:
          type: string
          description: |
            Mã lỗi nội bộ (xem DOC-04 mục 8.6).
            Ví dụ: `PDF_INVALID`, `OCR_RATE_LIMITED`, `DB_ERROR`, `INTERNAL_ERROR`.
        detail: { type: string }
        instance: { type: string }
        trace_id:
          type: string
          description: ID trace OpenTelemetry để debug
...
    PageKind:
      type: string
      enum: [native, scanned, hybrid]
...
    BBox:
      type: object
      description: Canonical Page Space (CPS) — `[x0,y0,x1,y1]` chuẩn hóa 0..1
      required: [x0, y0, x1, y1]
      properties:
        x0: { type: number, minimum: 0, maximum: 1 }
        y0: { type: number, minimum: 0, maximum: 1 }
        x1: { type: number, minimum: 0, maximum: 1 }
        y1: { type: number, minimum: 0, maximum: 1 }
...
    BBoxSource:
      type: string
      enum: [native, detector, estimated, human]
...
    Page:
      type: object
      properties:
        id: { type: string }
        document_id: { type: string }
        page_no: { type: integer }
        kind: { $ref: '#/components/schemas/PageKind' }
        width_pt: { type: number }
        height_pt: { type: number }
        rotation: { type: integer, enum: [0, 90, 180, 270] }
        render_dpi: { type: integer }
        preview_uri: { type: string, nullable: true }
        render_uri: { type: string, nullable: true }
        quality:
          type: object
          additionalProperties: true
          description: 'Chỉ số chất lượng: blur_score, skew_deg, contrast, ...'
...
    ClauseNode:
      type: object
      properties:
        id: { type: string }
        document_id: { type: string }
        node_type: { type: string, enum: [annex, article, clause, point, preamble, signature_block] }
        label: { type: string, description: 'VD: "Điều 5"' }
        number: { type: string, nullable: true }
        title: { type: string, nullable: true }
        text: { type: string }
        lang: { type: string, nullable: true }
        parent_id: { type: string, nullable: true }
        stable_path: { type: string, description: 'VD: "annex-01/art-5/cl-2/pt-a"' }
        page_start: { type: integer }
        page_end: { type: integer }
        confidence: { type: number }
        regions:
          type: array
          items:
            type: object
            properties:
              page_no: { type: integer }
              bbox: { $ref: '#/components/schemas/BBox' }
              bbox_source: { $ref: '#/components/schemas/BBoxSource' }
        children:
          type: array
          items: { $ref: '#/components/schemas/ClauseNode' }

    # ── Tables ─────────────────────────────────────────────────────────────
    TableCell:
      type: object
      properties:
        row_idx: { type: integer }
        col_idx: { type: integer }
        row_span: { type: integer, default: 1 }
        col_span: { type: integer, default: 1 }
        text: { type: string }
        bbox: { $ref: '#/components/schemas/BBox' }
        is_header: { type: boolean }
        confidence: { type: number }
    DocTable:
      type: object
      properties:
        id: { type: string }
        document_id: { type: string }
        page_no: { type: integer }
        bbox: { $ref: '#/components/schemas/BBox' }
        rows_count: { type: integer }
        cols_count: { type: integer }
        has_borders: { type: boolean }
        is_multi_page: { type: boolean }
        continued_from: { type: string, nullable: true }
        cells:
          type: array
          items: { $ref: '#/components/schemas/TableCell' }

    # ── Citation ───────────────────────────────────────────────────────────
...
    CitationResolved:
      type: object
      description: Output của GET /citations/{id} — đủ thông tin để UI highlight một thao tác
      properties:
        citation: { $ref: '#/components/schemas/Citation' }
        page_images:
          type: array
          description: Một URL ảnh preview cho mỗi trang có segment
          items:
            type: object
            properties:
              page_no: { type: integer }
              preview_url: { type: string }

    # ── Fact ───────────────────────────────────────────────────────────────
...
    FactEffective:
      type: object
      description: Fact kèm giá trị hiệu lực sau review + ETag cho optimistic concurrency
      properties:
        fact: { $ref: '#/components/schemas/Fact' }
        machine_value: {}
        effective_value: { nullable: true }
        review_state:
          type: string
          enum: [unreviewed, confirmed, corrected, rejected, awaiting_evidence]
        reviewer_id: { type: string, nullable: true }
        reviewed_at: { type: string, format: date-time, nullable: true }
        review_item_id:
          type: string
          nullable: true
          description: ID của review_item tương ứng (NULL = chưa sinh item)
        current_version:
          type: integer
          minimum: 0
          description: |
            Version hiện tại của review_item (0 = giá trị máy, 1..N = sau N lần action).
            **Client PHẢI gửi lại giá trị này trong `base_version` khi submit action.**

    # ── Finding ────────────────────────────────────────────────────────────
...
    ApiEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          nullable: true
        meta:
          type: object
          properties:
            request_id: { type: string }
            trace_id: { type: string }
            timestamp: { type: string, format: date-time }
    Pagination:
      type: object
      properties:
        next_cursor: { type: string, nullable: true }
        has_more: { type: boolean }
        total:
          type: integer
          nullable: true
          description: Có thể null nếu không đếm được (cursor-based)
    ApiEnvelopeDossierCreated:
      type: object
      properties:
        data:
          type: object
          properties:
            dossier_id: { type: string }
            job_id: { type: string }
        meta: { $ref: '#/components/schemas/ApiEnvelope/properties/meta' }
...
    ApiEnvelopePageList:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/Page' }
...
    ApiEnvelopeClauseTree:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/ClauseNode' }
    ApiEnvelopeTableList:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/DocTable' }
    ApiEnvelopeFactList:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/FactEffective' }
            pagination: { $ref: '#/components/schemas/Pagination' }
...
    ApiEnvelopeCitationResolved:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/CitationResolved' }
...
    CreateRunRequest:
      type: object
      properties:
        config_override:
          type: object
          properties:
            ocr_profile: { type: string, enum: [standard, high_res_binarize, table_optimized] }
            prompt_candidate_id: { type: string }
            confidence_threshold: { type: number, minimum: 0, maximum: 1 }
    PipelineRunSummary:
      type: object
      required: [run_id, dossier_id, status, created_at]
      properties:
        run_id: { type: string }
        dossier_id: { type: string }
        status: { type: string, enum: [queued, running, completed, failed, cancelled] }
        triggered_by: { type: string }
        duration_seconds: { type: number, nullable: true }
        created_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time, nullable: true }
    PipelineRunDetail:
      allOf:
        - $ref: '#/components/schemas/PipelineRunSummary'
        - type: object
          properties:
            stages:
              type: array
              items:
                type: object
                required: [stage_name, status]
                properties:
                  stage_name: { type: string }
                  status: { type: string, enum: [pending, running, completed, failed, skipped] }
                  duration_seconds: { type: number, nullable: true }
                  error_message: { type: string, nullable: true }
            metrics:
              type: object
              properties:
                pages_processed: { type: integer }
                facts_extracted: { type: integer }
                conflicts_found: { type: integer }
                tokens_consumed: { type: integer }
                cost_estimate_usd: { type: number }

    # ── Batch Schemas ───────────────────────────────────────────────────────
...
    ApiEnvelopePipelineRun:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/PipelineRunSummary' }
    ApiEnvelopePipelineRunList:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/PipelineRunSummary' }
            pagination: { $ref: '#/components/schemas/Pagination' }
    ApiEnvelopePipelineRunDetail:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/PipelineRunDetail' }
...
  responses:
    Unauthorized:
      description: Token JWT không hợp lệ, đã hết hạn hoặc không được cung cấp (401)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Không có quyền truy cập theo vai trò x-rbac hoặc không khớp X-Tenant-Id (403)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    BadRequest:
      description: Request không hợp lệ (validation, schema)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource không tồn tại
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token chuẩn RFC 7519 chứa các claims:
        - `sub`: User ID (ULID)
        - `tenant_id`: Định danh Tenant (bắt buộc khớp với header X-Tenant-Id)
        - `role`: OPERATOR | REVIEWER | ADMINISTRATOR
    tenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-Id
      description: |
        Header định danh Tenant phục vụ phân lập đa khách hàng (Tenant Isolation).
        Bắt buộc phải trùng khớp với claim `tenant_id` trong Bearer JWT.

# ═══════════════════════════════════════════════════════════════════════════
# SECURITY, WEBHOOKS, CALLBACKS
# ═══════════════════════════════════════════════════════════════════════════
security:
  - bearerAuth: []
    tenantHeader: []
...